chore: cleaup, remove unsed store, move files and removed consol.log
