--- src/stores/mappings.ts
@@ -1,5 +1,0 @@
-import { writable } from "svelte/store";
-
-export const siteToDefaultCollectionIdStore = writable<Map<string, string>>(
-    new Map(),
-);
